requirements.txt
gym==0.20.0
mujoco==2.3.0
mujoco-py==1.50.1.0
numpy==1.20.1
pytorch==1.10.0